server/router: Add global access policy
This allows to set an access policy similar to router::add_policy() but instead for the entire router rather than specific endpoints.

Theoretically, we could just use the existing router::add_policy() function and use a regex pattern matching all endpoints (such as .+) but this is error prone as we simply look for the first matching policy when checking policies added via router::add_policy().