server/router: improve the way policies are handled
This changes the way that policies are handled. Basically we're performing the policy check as the very first thing inside router::handle_request(). This ensures that we can also apply policies to entire sub-routers. Prior to this change, policies were only checked/handled within router::handle_http_request() which meant that sub-routers could not be policy protected. As a workaround, router::set_policy() was introduced to provide a router-wide policy.
This change fixes all of those problems. One can now simply use router::add_policy() and it will work for sub-router protection as well.

Ref #96