fix: accidental removal of m_workguard from controller_run_result
This actually was informative because it showed how m_workguard actually
does matter. See if there is no work guard, and the user calls
run() on a _client_ controller, then enough time elapses so that
all work is complete and there is a pause before new work is added,
the io_context::run() will return for each of the threads, meaning that
in actual fact queuing more actions via the controller (which is valid
after calling start) would not actually do anything.

tl;dr nasty obscure bug, keep the workguard its needed for clients