Make server::router not move-able.
Routes added to the router have a raw reference to the router itself.
When a router is moved, even though all the members are moved, the
raw references would not update themselves, which makes them dangling.

Downgrade router from move-able to non-move-able, and make
routing_context store a unique_ptr to router instead of an object.
This way, moving the routing_context would not end up creating a
new router, so references stored in the routes would remain valid.